// webapp/src/api/model/species.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { SpeciesAdvantageTraitsItem } from './speciesAdvantageTraitsItem';
import type { SpeciesDisadvantageTraitsItem } from './speciesDisadvantageTraitsItem';
import type { Nation } from './nation';

export interface Species {
  advantageTraits: SpeciesAdvantageTraitsItem[];
  description: string;
  disadvantageTraits: SpeciesDisadvantageTraitsItem[];
  id?: string;
  name: string;
  nations: Nation[];
  playable?: boolean;
}